Fix trunk on various platforms.
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@9588 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_player.cpp
@@ -5261,7 +5261,7 @@ bool player_t::parse_talents_armory( const std::string& talent_string )
   int encoding[ MAX_TALENT_SLOTS ];
 
   unsigned int i, i_max = std::min( talent_string.size(),
-                                    static_cast<unsigned>( MAX_TALENT_SLOTS ) );
+                                    static_cast< size_t >( MAX_TALENT_SLOTS ) );
   for ( i = 0; i < i_max; i++ )
   {
     char c = talent_string[ i ];
